split proxy/sso_test.go
